move client and server constants into const folder
