+++ islands/DateChanger.tsx
@@ -1,4 +1,4 @@
-import { MONTH_NUMBER_STRING } from "@/src/client_constants.ts";
+import { MONTH_NUMBER_STRING } from "@/src/const/client_constants.ts";
 
 enum Direction {
   PREV = 1,
