+++ routes/auth/login.tsx
@@ -1,6 +1,6 @@
 import { HandlerContext, Handlers } from "$fresh/server.ts";
 import { Head } from "$fresh/runtime.ts";
-import { DEPLOYMENT_ID } from "@/src/server_constants.ts";
+import { DEPLOYMENT_ID } from "@/src/const/server_constants.ts";
 
 export async function handler(
   req: Request,
