+++ routes/calendar/[year]/[month].tsx
@@ -3,7 +3,7 @@ import { getCookies } from "$std/http/cookie.ts";
 import {
   AUDIO_LOGBOOK_AUTH_COOKIE_NAME,
   DEPLOYMENT_ID,
-} from "@/src/server_constants.ts";
+} from "@/src/const/server_constants.ts";
 import { dbPromise } from "@/src/db/db.ts";
 import { UserSession } from "@/src/db/db_schema.ts";
 import { secretsPromise } from "@/src/secrets.ts";
